Worker process should also be waited
There are cases (such as the port being already in use or other fatal errors) when the server closes unexpectedly and the worker process is left running in the background with no easy option to catch it.

I changed the run command, such that it also waits for the worker. I have also add some red messages to point the die event of the server process.